refactor: sperated resource filter badges into seperate file BadgesList
