+++ src/Components/Resource/ListView.tsx
@@ -16,6 +16,7 @@ import Pagination from "../Common/Pagination";
 import { Modal, Button } from "@material-ui/core";
 
 import { limit, formatFilter } from "./Commons";
+import BadgesList from "./BadgesList";
 
 const Loading = loadable(() => import("../Common/Loading"));
 const PageTitle = loadable(() => import("../Common/PageTitle"));
@@ -75,6 +76,9 @@ export default function ListView() {
   };
 
   const appliedFilters = formatFilter(qParams);
+  const updateFilter = (params: any) => {
+    updateQuery(params);
+  };
 
   const refreshList = () => {
     fetchData();
@@ -112,29 +116,6 @@ export default function ListView() {
     qParams.ordering,
     offset,
   ]);
-
-  const removeFilter = (paramKey: any) => {
-    const params = { ...qParams };
-    console.log(params);
-    params[paramKey] = "";
-    updateQuery(params);
-  };
-
-  const badge = (key: string, value: any, paramKey: string) => {
-    return (
-      value && (
-        <span className="inline-flex items-center px-3 py-1 rounded-full text-xs font-medium leading-4 bg-white text-gray-600 border">
-          {key}
-          {": "}
-          {value}
-          <i
-            className="fas fa-times ml-2 rounded-full cursor-pointer hover:bg-gray-500 px-1 py-0.5"
-            onClick={(e) => removeFilter(paramKey)}
-          ></i>
-        </span>
-      )
-    );
-  };
 
   let showResourceCardList = (data: any) => {
     if (data && !data.length) {
@@ -281,58 +262,7 @@ export default function ListView() {
         </div>
       </div>
 
-      <div className="flex flex-wrap space-x-2 mt-2 ml-2 space-y-1">
-        {badge("Ordering", appliedFilters.ordering, "ordering")}
-        {badge(
-          "status",
-          appliedFilters.status != "--" && appliedFilters.status,
-          "status"
-        )}
-        {badge(
-          "Emergency",
-          appliedFilters.emergency === "true"
-            ? "yes"
-            : appliedFilters.emergency === "false"
-            ? "no"
-            : undefined,
-          "emergency"
-        )}
-        {badge(
-          "Modified After",
-          appliedFilters.modified_date_after,
-          "modified_date_after"
-        )}
-        {badge(
-          "Modified Before",
-          appliedFilters.modified_date_before,
-          "modified_date_before"
-        )}
-        {badge(
-          "Created Before",
-          appliedFilters.created_date_before,
-          "created_date_before"
-        )}
-        {badge(
-          "Created After",
-          appliedFilters.created_date_after,
-          "created_date_after"
-        )}
-        {badge(
-          "Filtered By",
-          appliedFilters.assigned_facility && "Assigned Facility",
-          "assigned_facility"
-        )}
-        {badge(
-          "Filtered By",
-          appliedFilters.orgin_facility && "Origin Facility",
-          "orgin_facility"
-        )}
-        {badge(
-          "Filtered By",
-          appliedFilters.approving_facility && "Resource Approving Facility",
-          "approving_facility"
-        )}
-      </div>
+      <BadgesList appliedFilters={appliedFilters} updateFilter={updateFilter} />
 
       <div className="px-4">
         {isLoading ? (
